Cleanup example beans
