+++ elide-graphql/README.md
@@ -1,0 +1,4 @@
+elide-graphql
+=============
+
+[![Dependency Status](https://www.versioneye.com/user/projects/5621503e36d0ab0016000a41/badge.svg?style=flat)](https://www.versioneye.com/user/projects/5621503e36d0ab0016000a41)
