+++ elide-graphql/src/test/java/com/yahoo/elide/graphql/PersistentResourceFetcherTest.java
@@ -9,13 +9,13 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.yahoo.elide.ElideSettings;
 import com.yahoo.elide.ElideSettingsBuilder;
-import com.yahoo.elide.books.Author;
-import com.yahoo.elide.books.Book;
-import com.yahoo.elide.books.Pseudonym;
-import com.yahoo.elide.books.Publisher;
 import com.yahoo.elide.core.datastore.inmemory.InMemoryDataStore;
 import com.yahoo.elide.core.datastore.inmemory.InMemoryTransaction;
 import com.yahoo.elide.core.filter.dialect.RSQLFilterDialect;
+import example.Author;
+import example.Book;
+import example.Pseudonym;
+import example.Publisher;
 import graphql.ExecutionResult;
 import graphql.GraphQL;
 import graphql.GraphQLError;
