+++ elide-graphql/src/test/java/example/Address.java
@@ -3,8 +3,7 @@
  * Licensed under the Apache License, Version 2.0
  * See LICENSE file in project root for terms.
  */
-
-package com.yahoo.elide.books;
+package example;
 
 import com.yahoo.elide.annotation.Include;
 import lombok.Getter;
@@ -15,9 +14,9 @@ import javax.persistence.Entity;
 @Include
 @Entity
 public class Address {
-
     @Getter @Setter
     private String street1;
+
     @Getter @Setter
     private String street2;
