+++ elide-graphql/src/test/java/example/Book.java
@@ -3,7 +3,7 @@
  * Licensed under the Apache License, Version 2.0
  * See LICENSE file in project root for terms.
  */
-package com.yahoo.elide.books;
+package example;
 
 import com.yahoo.elide.annotation.Audit;
 import com.yahoo.elide.annotation.Include;
@@ -22,6 +22,8 @@ import com.yahoo.elide.annotation.OnUpdatePreSecurity;
 import com.yahoo.elide.annotation.SharePermission;
 import com.yahoo.elide.security.RequestScope;
 
+import java.util.ArrayList;
+import java.util.Collection;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -29,8 +31,6 @@ import javax.persistence.Id;
 import javax.persistence.ManyToMany;
 import javax.persistence.OneToOne;
 import javax.persistence.Table;
-import java.util.ArrayList;
-import java.util.Collection;
 
 /**
  * Model for books.
@@ -171,8 +171,8 @@ public class Book {
         // book being read post commit
     }
 
-    @Override
-    public String toString() {
-        return "Book{id=" + id + ", title='" + title + "\'}";
+    @OnUpdatePreCommit
+    public void alwaysOnUpdate() {
+        // should be called on _any_ update
     }
 }
