+++ elide-graphql/src/test/java/example/Pseudonym.java
@@ -3,52 +3,43 @@
  * Licensed under the Apache License, Version 2.0
  * See LICENSE file in project root for terms.
  */
-
-package com.yahoo.elide.books;
+package example;
 
 import com.yahoo.elide.annotation.Include;
+import lombok.Getter;
+import lombok.Setter;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import java.util.HashSet;
-import java.util.Set;
+import javax.persistence.OneToOne;
 
 /**
- * Model for publisher.
+ * An author's pen-name
  */
 @Entity
 @Include
-public class Publisher {
+public class Pseudonym {
     private long id;
+    private Author author;
+
+    @Getter @Setter
     private String name;
-    private Set<Book> books = new HashSet<>();
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     public long getId() {
         return id;
     }
-
     public void setId(long id) {
         this.id = id;
     }
 
-    public String getName() {
-        return name;
+    @OneToOne
+    public Author getAuthor() {
+        return author;
     }
-
-    public void setName(String title) {
-        this.name = title;
-    }
-
-    @OneToMany(mappedBy = "publisher")
-    public Set<Book> getBooks() {
-        return books;
-    }
-
-    public void setBooks(Set<Book> books) {
-        this.books = books;
+    public void setAuthor(Author author) {
+        this.author = author;
     }
 }
