+++ elide-graphql/src/test/java/example/Readme.md
@@ -1,0 +1,4 @@
+Example bean annotations
+=
+
+Sample Parent and Child JPA annotated beans with additional Elide annotations.
