+++ elide-graphql/src/test/java/example/package-info.java
@@ -1,0 +1,6 @@
+/*
+ * Copyright 2017, Yahoo Inc.
+ * Licensed under the Apache License, Version 2.0
+ * See LICENSE file in project root for terms.
+ */
+package example;
